feat: aggressive SEO improvements across all routes
Home, /tours and /personalize had no page-level metadata and shared
the same default title/description, competing with each other in
search results. Adds unique title/description/canonical per route,
an ItemList schema for the tours listing, an H1 with indexable copy
on the home page, canonical URLs on tour detail pages, the Google
Search Console verification tag, and lastModified/host to the
sitemap and robots files.
