+++ src/app/sitemap.ts
@@ -4,11 +4,19 @@ import { MetadataRoute } from 'next'
 const BASE_URL = 'https://roadmapcol.com'
 
 export default function sitemap(): MetadataRoute.Sitemap {
+  const lastModified = new Date()
+
   const staticRoutes: MetadataRoute.Sitemap = [
-    { changeFrequency: 'weekly', priority: 1, url: BASE_URL },
-    { changeFrequency: 'weekly', priority: 0.9, url: `${BASE_URL}/tours` },
+    { changeFrequency: 'weekly', lastModified, priority: 1, url: BASE_URL },
+    {
+      changeFrequency: 'weekly',
+      lastModified,
+      priority: 0.9,
+      url: `${BASE_URL}/tours`,
+    },
     {
       changeFrequency: 'monthly',
+      lastModified,
       priority: 0.7,
       url: `${BASE_URL}/personalize`,
     },
@@ -17,6 +25,7 @@ export default function sitemap(): MetadataRoute.Sitemap {
   const uniqueTourHrefs = [...new Set(TOURS.map((tour) => tour.href))]
   const tourRoutes: MetadataRoute.Sitemap = uniqueTourHrefs.map((href) => ({
     changeFrequency: 'monthly',
+    lastModified,
     priority: 0.8,
     url: `${BASE_URL}${href}`,
   }))
