+++ src/app/tours/__tests__/tour-page.test.tsx
@@ -59,6 +59,7 @@ describe('generateMetadata', () => {
     })
     expect(result.title).toBe('Test Tour')
     expect(result.description).toBe('Test description')
+    expect(result.alternates?.canonical).toBe('/tours/test-tour')
   })
 
   it('returns empty object for an unknown tour', async () => {
